fix: made type-metadata.storage.ts much faster
Changed the data modeling of type-metadata storage to store the data in maps instead of arrays.
This results in a dramatic improvement of application startup time since many of the lookups for metadata are now in O(1) instead of O(n)
